+++ node/queries/getUser.ts
@@ -1,0 +1,18 @@
+import { print } from 'graphql'
+import gql from 'graphql-tag'
+
+export default print(gql`
+  query user($id: ID!) {
+    getUser(id: $id) {
+      id
+      roleId
+      userId
+      clId
+      orgId
+      costId
+      name
+      email
+      canImpersonate
+    }
+  }
+`)