+++ node/queries/listRoles.ts
@@ -1,0 +1,12 @@
+import { print } from 'graphql'
+import gql from 'graphql-tag'
+
+export default print(gql`
+  query roles {
+    listRoles {
+      id
+      name
+      slug
+    }
+  }
+`)